introduce test data for serialization
The `data` subproject includes a `telemetry.csv` file with vehicle
telemetry data for a trip from Tennessee to Indiana.

Since I don't have a fancy telemetry collection system, we're going
to fake it with GPS data captured during a trip from Alabama to
Indiana earlier this year.  The `Telemetry` object has a `main()`
method that takes CSV data from GPS measurements and augments it
with distance and location values.

It would be nice to include some `byte` values to represent sensor
data, but I have not done that yet.  And the current data is a
reasonable place to start testing serialization libraries.  First
up with be JSON with [circe].

[circe]: https://circe.github.io/circe/